requirements.txt
Django==2.2.19
pytz==2021.3
sqlparse==0.4.2
djangorestframework==3.12.4
drf-yasg==1.20.0